requirements.txt
streamlit
boto3
confluent-kafka
PyGithub
requests
sqlalchemy
python-dotenv
google-generativeai
